aula 15/03 = elif

+++ programas/exercicios03/exercicio01.py
@@ -1,5 +1,5 @@
-valor_produto = int(input("Quanto custa o produto que você quer comprar ? "))
-valor_especie = int(input("Quanto você tem em mãos?  "))
+valor_produto = float(input("Quanto custa o produto que você quer comprar ? "))
+valor_especie = float(input("Quanto você tem em mãos?  "))
 
 if (valor_especie >= valor_produto ):
     print("Você pode comprar o produto")
